requirements_bk.txt
fastapi==0.120.3
uvicorn==0.38.0
pydantic==2.12.3
email-validator==2.3.0
python-multipart==0.0.20
sqlalchemy==2.0.44
pymysql==1.1.2
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
passlib==1.7.4
pillow==12.0.0
torch==2.2.0
torchvision==0.17.0
starlette==0.49.1